+++ packages/at_client/lib/src/at_collection/impl/at_collection_operations_impl.dart
@@ -6,6 +6,7 @@ import 'package:at_client/src/at_collection/collection_util.dart';
 import 'package:at_utils/at_logger.dart';
 import 'collection_methods_impl.dart';
 
+@Deprecated("Use AtClient.collection for collection-style operations")
 class AtCollectionModelOperationsImpl implements AtCollectionModelOperations {
   final _logger = AtSignLogger('AtCollectionModelOperationsImpl');
   late AtCollectionModel atCollectionModel;