+++ packages/at_client/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 3.13.0
+
+Several significant enhancements to the API to make it much easier to use.
+- feat: New feature - Collections - a clean API for storing, sharing, 
+  unsharing and deleting objects in named collections, with sub-collections, 
+  event streams, built-in support for read receipts, and more
+- feat: added a new method, `send`, to NotificationService which is much 
+  easier to use than the old (still fine to use) `notify` method.
+- feat: added `factory AtRpc.server` to make it much simpler to create AtRpc 
+  servers. 
+- docs: Rewrote the main README
+- docs: Added many examples in the [example](example/README.md) directory
+
 ## 3.12.0
 
 - feat: explicit AtClient lifecycle control — cleanly stop and resume atClients without
@@ -62,7 +75,7 @@
 
 ## 3.6.0
 - feat: deprecate the (misleadingly named)
-  `AtClientPreference.atProtocolEmitted` and change its default value
+  `AtClientPreference.Atsign ProtocolEmitted` and change its default value
   from 1.5.0 to 2.0.0
 
 ## 3.5.3
@@ -206,7 +219,7 @@
 ## 3.0.61
 - fix: ensure key exchange functions properly when the sync service is not
   being used
-- feat: Add AtRpc - A simple rpc request-response API which uses atProtocol
+- feat: Add AtRpc - A simple rpc request-response API which uses Atsign Protocol
   notifications under the hood.
 ## 3.0.60
 - feat: Add `useRemoteAtServer` to PutRequestOptions. When set, the update